server/requirements.txt
flask==2.2.3
flask_cors==3.0.10
matplotlib==3.7.1
torchvision==0.15.1
opencv_python==4.7.0.72
tensorboard==2.12.0
pyngrok==5.2.1